- Changed instances of integer to long (or decimal in case of FHIR) to support larger resource generation than Integer.MAX_VALUE bytes - Removed hard-coded resource size limit of 100 MB

+++ src/main/java/dev/dsf/bpe/ConstantsPing.java
@@ -146,7 +146,7 @@ public final class ConstantsPing
 
 	public static final String PONG_ERROR_MESSAGE_CLEANUP_TIMEOUT = "Timeout while waiting for cleanup message";
 
-	public static final int DOWNLOAD_RESOURCE_SIZE_BYTES_DEFAULT = 10000000;
+	public static final long DOWNLOAD_RESOURCE_SIZE_BYTES_DEFAULT = 10000000L;
 
 	public static final MediaType DOWNLOAD_RESOURCE_MIME_TYPE = MediaType.APPLICATION_OCTET_STREAM_TYPE;
