+++ src/main/java/dev/dsf/bpe/util/task/input/generator/DownloadResourceSizeGenerator.java
@@ -1,7 +1,7 @@
 package dev.dsf.bpe.util.task.input.generator;
 
 import org.hl7.fhir.r4.model.Coding;
-import org.hl7.fhir.r4.model.IntegerType;
+import org.hl7.fhir.r4.model.DecimalType;
 import org.hl7.fhir.r4.model.Task;
 
 import dev.dsf.bpe.ConstantsPing;
@@ -12,10 +12,10 @@ public final class DownloadResourceSizeGenerator
 	{
 	}
 
-	public static Task.ParameterComponent create(int sizeBytes)
+	public static Task.ParameterComponent create(long sizeBytes)
 	{
 		Task.ParameterComponent param = new Task.ParameterComponent();
-		param.setValue(new IntegerType(sizeBytes)).getType().addCoding(new Coding(ConstantsPing.CODESYSTEM_DSF_PING,
+		param.setValue(new DecimalType(sizeBytes)).getType().addCoding(new Coding(ConstantsPing.CODESYSTEM_DSF_PING,
 				ConstantsPing.CODESYSTEM_DSF_PING_VALUE_DOWNLOAD_RESOURCE_SIZE_BYTES, null));
 		return param;
 	}
